sideBarFooter 작업 및 헤더 삭제
