--- src/components/Footer.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import "../styles/Footer.css";
-
-const Footer = () => {
-    return (
-        <footer className="footer">
-            © 2025 Grou-up. All Rights Reserved.
-        </footer>
-    );
-};
-
-export default Footer;
